requirements.txt
# /Users/radiant/Desktop/RXinDexer/requirements.txt
# This file contains all the dependencies required for the RXinDexer project.
# It lists specific versions to ensure compatibility across development and production environments.

# Radiant Node RPC Communication
python-bitcoinrpc==1.0

# CBOR Parsing for Glyph Tokens
cbor2==5.4.6

# Web API Framework
fastapi==0.115.0
uvicorn==0.27.1
pydantic==2.6.1

# Database ORM and Connection
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.8

# Testing
pytest==8.3.3
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1

# Bloom Filter for Performance Optimization (Optional)
pybloom-live==4.0.0
